# Dart - make sure added padding is zeroed, same as in C++