I've noticed (especially with custom allocators and buffer reuse) that padding in the buffer isn't zeroed out, as it is in c++ (using `fill()`) - this PR aligns the behavior between dart an c++.