@vaind 2021-07-08
This applies after `reset()` (builder reuse) as well, regardless of the
allocator.

On Thu, Jul 8, 2021, 10:14 PM Dan Field ***@***.***> wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In dart/lib/flat_buffers.dart
> <https://github.com/google/flatbuffers/pull/6716#discussion_r666488980>:
>
> > +    // zero out the added padding
> +    for (var i = sizeBeforePadding + 1;
> +        i <= finishedSize - requiredBytes;
> +        i++) {
> +      _setUint8AtTail(_buf, i, 0);
> +    }
>
> Dart zero initializes ByteBuffers by default. I don't think this is the
> right place to do this. We should just document that custom allocators need
> to do this, no?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/google/flatbuffers/pull/6716#pullrequestreview-702461770>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABQOG4TY6TSMIFC3HK2QXP3TWYBLLANCNFSM47UH73GQ>
> .
>