@dnfield 2021-07-08
Can the `reset` method let the allocator know it needs to reset then? Some allocator implementations (like the default one, or a `calloc` based one) won't need to do this.