@vaind 2021-07-09
If you reuse the same builder, calling reset() between builds, you end up with arbitrary data (left over from the previous object) in the padding space. `reset()` has no chance to do anything about it, if we don't want to zero-out the whole of the previously used buffer, because at the point where you call `reset()`, you don't yet know where the padding will be