@dnfield 2021-07-09
Well, for the default implementation, it should be much cheaper for `reset` to just zero out the buffer if it is reusing it (and, specifically, to just new up a ByteBuffer rather than actually setting the data to zero - since it would involve only a single jump to C++ code rather than repeatedly jumping to it in a loop). I know that was changed in one of these patches recently, but if the price of changing it is this I strongly suspect it's not a great tradeoff.

It would probably help if we had benchmarks around this stuff, but setting up infra for that is not going to be fun :(. Do you have any measurements around this locally you're using?