@vaind 2021-07-09
> Benchmarks

Yes, actually I have some setup to benchmark the builder and reader performance, see  https://github.com/objectbox/flatbuffers-benchmark

I've run this specific scenario for the three situations discussed:
* 256 byte initial size (as used in the linked benchmark)
  * master branch: 2.618976844316231 us
  * zero padding branch: 2.9047152218620376 us
  * new buffer on reset: 2.830360984059415 us

* default 1024 byte initial size (only affects "new buffer on reset"):
  * new buffer on reset: 3.346755489512947 us

where "new buffer on reset" is this, applied to master:
```diff
Index: dart/lib/flat_buffers.dart
<+>UTF-8
===================================================================
diff --git a/dart/lib/flat_buffers.dart b/dart/lib/flat_buffers.dart
--- a/dart/lib/flat_buffers.dart	(revision bf3470c1610458eae0882b74903760a6f8231856)
+++ b/dart/lib/flat_buffers.dart	(date 1625817306896)
@@ -447,6 +447,8 @@
 
   /// Reset the builder and make it ready for filling a new buffer.
   void reset() {
+    _allocator.deallocate(_buf);
+    _buf = _allocator.allocate(initialSize);
     _maxAlign = 1;
     _tail = 0;
     _currentVTable = null;
```